Rename ContextAwareTokenCredential to OidcTokenCredential and move it out of the runtime package
