+++ extensions/oidc/runtime/src/main/java/io/quarkus/oidc/OidcTokenCredential.java
@@ -1,18 +1,18 @@
-package io.quarkus.oidc.runtime;
+package io.quarkus.oidc;
 
 import io.quarkus.security.credential.TokenCredential;
 import io.vertx.ext.web.RoutingContext;
 
-public class ContextAwareTokenCredential extends TokenCredential {
+public class OidcTokenCredential extends TokenCredential {
 
     private RoutingContext context;
 
-    protected ContextAwareTokenCredential(String token, String type, RoutingContext context) {
+    protected OidcTokenCredential(String token, String type, RoutingContext context) {
         super(token, type);
         this.context = context;
     }
 
-    RoutingContext getContext() {
+    public RoutingContext getRoutingContext() {
         return context;
     }
 }
