+++ integration-tests/oidc-code-flow/src/main/java/io/quarkus/it/keycloak/CustomSecurityIdentityAugmentor.java
@@ -1,0 +1,32 @@
+package io.quarkus.it.keycloak;
+
+import javax.enterprise.context.ApplicationScoped;
+
+import io.quarkus.oidc.IdTokenCredential;
+import io.quarkus.security.identity.AuthenticationRequestContext;
+import io.quarkus.security.identity.SecurityIdentity;
+import io.quarkus.security.identity.SecurityIdentityAugmentor;
+import io.quarkus.security.runtime.QuarkusSecurityIdentity;
+import io.smallrye.mutiny.Uni;
+import io.vertx.ext.web.RoutingContext;
+
+@ApplicationScoped
+public class CustomSecurityIdentityAugmentor implements SecurityIdentityAugmentor {
+
+    @Override
+    public int priority() {
+        return 0;
+    }
+
+    @Override
+    public Uni<SecurityIdentity> augment(SecurityIdentity identity, AuthenticationRequestContext context) {
+        IdTokenCredential cred = identity.getCredential(IdTokenCredential.class);
+        if (cred != null) {
+            QuarkusSecurityIdentity.Builder builder = QuarkusSecurityIdentity.builder(identity);
+            builder.addAttribute(RoutingContext.class.getName(), cred.getRoutingContext());
+            identity = builder.build();
+        }
+        return Uni.createFrom().item(identity);
+    }
+
+}
